[QOLDEV-490] extend job timeout if type guessing is needed
- Also refactor: move the function to check if a datastore entry exists, from the job module to utils,
so different modules can check it.
